requirements.txt
google-generativeai
gradio==5.0.0b3
pillow